move tf dependency to dev dep
