+++ README.md
@@ -6,3 +6,9 @@
 ![](https://img.shields.io/pypi/l/redshells)
 
 Machine learning tasks which are used with data pipeline library "luigi" and its wrapper "gokart".
+
+## Dependencies
+
+You should install `tensorflow = ">=1.13.1,<2.0"` manually if you would like to use some models.
+`Tensorflow` is not included redshells' dependencies because there are models not used it.
+For more details, see `redshells/model` directory.
